renaming cuda submodule and cleaning up the code
