+++ pyccel/cuda/.cph/.__init__.py_c1c278448be5f155f39d5932a658b053.prob
@@ -1,0 +1,1 @@
+{"name":"Local: __init__","url":"/home/said/Desktop/pyccel-cuda/pyccel/cuda/__init__.py","tests":[{"id":1717795885693,"input":"","output":""}],"interactive":false,"memoryLimit":1024,"timeLimit":3000,"srcPath":"/home/said/Desktop/pyccel-cuda/pyccel/cuda/__init__.py","group":"local","local":true}
