+++ tests/pyccel/scripts/kernel/hello_kernel.py
@@ -1,4 +1,18 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
-from func import f
+from pyccel.decorators import kernel
+from pyccel import cuda
+
+@kernel
+def say_hello(its_morning : bool):
+    if(its_morning):
+        print("Hello and Good morning")
+    else:
+        print("Hello and Good afternoon")
+
+def f():
+    its_morning = True
+    say_hello[1,1](its_morning)
+    cuda.synchronize()
+
 f()
